pyproject.toml
[tool.poetry]
name = "vortex"
version = "0.0.1"
description = ""
authors = ["Matheus Henrique <eng.matheushenrique@proton.me>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
SQLAlchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
httpx = "^0.25.1"
pydantic-settings = "^2.1.0"
sqlalchemy-utils = "^0.41.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-mail = "^1.4.1"
pytest = "^7.4.3"
flake8 = "^6.1.0"
alembic = "^1.12.1"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"

[tool.poe.tasks.test]
cmd = 'docker compose run --rm --entrypoint "sh -c" backend "pytest ${test_options}"'
args = [{name = 'test_options', positional = true, default = ''}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
